pyproject.toml
[project]
name = "memlord"
version = "0.1.0"
description = "MCP server for storing and searching memory"
license = {file = "LICENSE"}
requires-python = "==3.12.*"
dependencies = [
    "asyncpg>=0.29.0",
    "alembic>=1.18.4",
    "authlib>=1.3.0",
    "fastapi[all]>=0.135.1",
    "fastmcp>=3.1.0",
    "onnxruntime>=1.20.0",
    "pgvector>=0.3.0",
    "pydantic-settings>=2.7.0",
    "dateparser>=1.2.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "tokenizers>=0.21.0",
    "greenlet>=3.3.2",
    "email-validator>=2.3.0",
    "bcrypt>=5.0.0",
]

[project.scripts]
memlord = "memlord.main:main"

[dependency-groups]
dev = [
    "alembic-autogen-check",
    "black",
    "httpx",
    "pytest==9.*",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "pytest-xdist",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memlord"]

[tool.coverage.run]
source = ["src/memlord"]
parallel = true
concurrency = ["thread", "greenlet"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | .venv
  | migrations
  | data
)
'''
